requirements.txt
# AI-Powered Chrome Automation System Requirements
# สร้างโดย WAWA - เร็วขึ้น 5 เท่า

# Core Dependencies
selenium==4.18.1
webdriver-manager==4.0.1
pyautogui==0.9.54
opencv-python==4.11.0.86
pillow==11.2.1
numpy==2.2.6

# AI & Machine Learning
openai==1.58.1
transformers==4.53.0
torch==2.7.1
torchvision==0.22.1
sentence-transformers==5.0.0

# Thai Language Processing
pythainlp==5.1.2
easyocr==1.7.2
thai-segmenter==0.4.2

# Utilities
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
aiofiles==23.2.1

# Image Processing
scikit-image==0.25.2
scipy==1.15.3

# Data Processing
pandas==2.2.2
# matplotlib==3.9.1  # Commented out - requires build tools
# seaborn==0.13.2    # Commented out - requires matplotlib

# Testing
pytest==8.2.2
pytest-html==4.1.1

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Supabase Integration
supabase==2.0.0
